(Main): #031 Reflect preset settings change on org chart card
